Add rebuildVolumeForAllMarkets() and migration to run it
rebuildVolumeForAllMarkets() was added to facilitate a recent volume fix which
required a full DB reset.

It can also be used in future unit testing to verify that incremental and
non-incremental methods of volume calculation produce the correct result.

Or as an ongoing sanity check: executing rebuildVolumeForAllMarkets() on a
production database should result in no changes.

Related to AugurProject/augur#289