Add frozen funds (#821)
This is a backwards-incompatible change to public API getUserTradingPositions().

Fixes AugurProject/augur#988

See docs in frozen-funds.ts. Adds new concept of "frozen funds", which are
tokens a user had to escrow to obtain their current position.

Frozen funds are tracked per market outcome, stored in a time series in the DB.

The latest frozen funds for each outcome are included in
getUserTradingPositions() response.

A grand total frozen funds is also included in getUserTradingPositions()
response, and, if the user is a market creator, includes escrowed ETH in market
validity bonds.