Add tradingPositionsTotal to getUserTradingPositions()
This is a backwards-incompatible change to public API getUserTradingPositions()

Fixes AugurProject/augur#1647

1. Adds GetUserTradingPositionsResponse.tradingPositionsTotal, a portfolio-level
aggregation of trading positions.

2. GetUserTradingPositionsResponse.{tradingPositionsTotal,frozenFundsTotal} are
now defined if and only if UserTradingPositionsParams.marketId is undefined.
Ie. the totals are provided if and only if request is _not_ filtered by market.
This is because data to compute totals is (locally) unavailable when filtering
by marketId.