Require LastTradePriceMinusMinPrice to be defined
In financial-math, LastTradePriceMinusMinPrice previously could be undefined due
to data unavailability. We'd then interpret undefined as zero at some point
during a specific calculation.

The problem with this design is 1. interpreting undefined as zero leads to
unexpected, calculation-specific results 2. in future, other inputs might be
undefined due to data unavailability, exacerbating (1). (Another price was
discovered to be potentially undefined.)

To fix this problem we have at least two design options

Solution 1. allow inputs to financial-math to be undefined, but require that
outputs become undefined if any input is. In this case the undefined outputs
become a problem downstream, because clients interpret undefined on a
case-by-case basis, which erodes the standardization we hope to achieve with
financial-math.

Solution 2. require all inputs to financial-math to be defined. In this case
clients must decide on a default value for inputs prior to calling into
financial-math, which localizes special-case handling of data unavailability.

Solution 2 was implemented here.