Fixes #1849
During self trades, there are two PL rows generated for one logIndex.
Since we weren't taking this into accoutn in the query, the natural
ordering of SQLite meant the `limit 1` was picking up the first half
of the trade via the first PL row.

Added ordering by rowid DESC in order to make sure the LAST trade
generates the PL we are using to determine how the positions change over
time.