Disable 20190205102050_rm_complete_sets_trade_group_id.ts (#872)
This migration dropped column completeSets.tradeGroupId

This migration always fails since commit cd2ceba73 which added
`numCompleteSets: true` to post-process-database-results. It always fails
because sqlite doesn't support drop column natively and Knex implements this
internally by using a temp table; that temp table is affected by
post-process-database-results because the completeSets.numCompleteSets column
is converted to BigNumber when Knex reads it and then that BigNumber fails to
be written into the temp table.

The fix is to make the migration a no-op without actually deleting it.
This allows clients who already ran the migration to still find the migration
file (and Knex checks for this), but clients who haven't yet run the migration
will succeed because it's a no-op.